refactor: local strategy
